refactor: rename styles on book details page
